requirements.txt
google-adk
python-multipart
pypdf2
python-docx
pydantic
dotenv
google-genai
google-generativeai
google-cloud-storage
IPython
Flask
weasyprint